pyproject.toml
[project]
name = "new-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=0.1.0",
    "apscheduler>=3.10.4",
    "colorlog>=6.8.2",
    "fastapi>=0.111.0",
    "finnhub-python>=2.4.24",
    "httpx>=0.27.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "mcp>=1.2.0",
    "numpy>=1.26.4",
    "openai>=1.37.0",
    "orjson>=3.10.6",
    "pandas>=2.2.2",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "qdrant-client[fastembed]>=1.9.1",
    "scipy>=1.13.1",
    "streamlit>=1.36.0",
    "tenacity>=8.3.0",
    "uvicorn>=0.30.1",
    "websockets>=15.0.1",
    "yfinance>=0.2.65",
]
